+++ plugins/cordova-sqlite-storage/spec/www/spec/db-tx-error-mapping-test.js
@@ -2,11 +2,20 @@
 
 var MYTIMEOUT = 12000;
 
-var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
+// NOTE: DEFAULT_SIZE wanted depends on type of browser
 
-var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
-var isWindows = /Windows /.test(navigator.userAgent); // Windows
+var isWindows = /MSAppHost/.test(navigator.userAgent);
 var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
+var isFirefox = /Firefox/.test(navigator.userAgent);
+var isWebKitBrowser = !isWindows && !isAndroid && /Safari/.test(navigator.userAgent);
+var isBrowser = isWebKitBrowser || isFirefox;
+var isEdgeBrowser = isBrowser && (/Edge/.test(navigator.userAgent));
+var isChromeBrowser = isBrowser && !isEdgeBrowser && (/Chrome/.test(navigator.userAgent));
+var isSafariBrowser = isWebKitBrowser && !isEdgeBrowser && !isChromeBrowser;
+
+// should avoid popups (Safari seems to count 2x)
+var DEFAULT_SIZE = isSafariBrowser ? 2000000 : 5000000;
+// FUTURE TBD: 50MB should be OK on Chrome and some other test browsers.
 
 // NOTE: While in certain version branches there is no difference between
 // the default Android implementation and implementation #2,
@@ -18,24 +27,26 @@ var scenarioList = [
   'Plugin-implementation-2'
 ];
 
-var scenarioCount = (!!window.hasWebKitBrowser) ? (isAndroid ? 3 : 2) : 1;
+var scenarioCount = (!!window.hasWebKitWebSQL) ? (isAndroid ? 3 : 2) : 1;
 
 var mytests = function() {
 
   for (var i=0; i<scenarioCount; ++i) {
+    // TBD skip plugin test on browser platform (not yet supported):
+    if (isBrowser && (i === 0)) continue;
 
-    // GENERAL: SKIP ALL on WP8 for now
-    describe(scenarioList[i] + ': db tx error mapping test(s)' +
-             ((isWindows && !isWP8) ?
-              ' [Windows version with INCORRECT error code (0) & INCONSISTENT error message (missing actual error info)]' :
-               ''), function() {
+    describe(scenarioList[i] + ': db tx error mapping test(s) [TBD INCORRECT & INCONSISTENT error message on Windows - missing actual error info ref: litehelpers/Cordova-sqlite-storage#539]', function() {
       var scenarioName = scenarioList[i];
       var suiteName = scenarioName + ': ';
       var isWebSql = (i === 1);
       var isImpl2 = (i === 2);
+      // XXX TBD WORKAROUND SOLUTION for (WebKit) Web SQL on Safari browser (TEST DB NAME IGNORED):
+      var recycleWebDatabase = null;
 
       // NOTE: MUST be defined in function scope, NOT outer scope:
       var openDatabase = function(name, ignored1, ignored2, ignored3) {
+        if (isWebSql && isSafariBrowser && !!recycleWebDatabase)
+          return recycleWebDatabase;
         if (isImpl2) {
           return window.sqlitePlugin.openDatabase({
             // prevent reuse of database from default db implementation:
@@ -46,7 +57,8 @@ var mytests = function() {
           });
         }
         if (isWebSql) {
-          return window.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
+          return recycleWebDatabase =
+            window.openDatabase(name, '1.0', 'Test', DEFAULT_SIZE);
         } else {
           return window.sqlitePlugin.openDatabase({name: name, location: 0});
         }
@@ -82,8 +94,6 @@ var mytests = function() {
         // GENERAL NOTE: ERROR MESSAGES are subject to improvements and other possible changes.
 
         it(suiteName + 'syntax error: command with misspelling', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // SKIP for now
-
           var db = openDatabase("Syntax-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -104,7 +114,7 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
               if (isWindows || (isAndroid && isImpl2))
@@ -112,13 +122,13 @@ var mytests = function() {
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 near \"SLCT\": syntax error/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*near \"SLCT\": syntax error/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/near \"SLCT\": syntax error.*code 1.*while compiling: SLCT 1/);
               else
                 expect(error.message).toMatch(/near \"SLCT\": syntax error/);
@@ -135,7 +145,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWindows || isWebSql || (isAndroid && isImpl2))
@@ -158,9 +168,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'INSERT syntax error [VALUES in the wrong place] with a trailing space [XXX "incomplete input" message]', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
+        it(suiteName + 'INSERT with VALUES in the wrong place (with a trailing space) [XXX TBD "incomplete input" vs "syntax error" message on (WebKit) Web SQL on Android 8.x/...]', function(done) {
           var db = openDatabase("INSERT-Syntax-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -184,7 +192,7 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
               if (isWindows || (isAndroid && isImpl2))
@@ -192,17 +200,22 @@ var mytests = function() {
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && (/Android [7-9]/.test(navigator.userAgent)))
+                expect(error.message).toMatch(/could not prepare statement.*/); // XXX TBD incomplete input vs syntax error message on Android 8(+)
+              else if (isWebSql && !isChromeBrowser && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 near \"VALUES\": syntax error/);
+              else if (isWebSql && isBrowser)
+                expect(error.message).toMatch(/could not prepare statement.*1 incomplete input/);
+              else if (isWebSql)
+                expect(error.message).toMatch(/near \"VALUES\": syntax error/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              //* else if (isAndroid && !isImpl2) //* XXX TBD Android (default implementation) vs ...
-              //*   expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*near \" \": syntax error/);
+              else if (isAndroid && !isImpl2)
+                expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*incomplete input/);
               else if (isAndroid && isImpl2)
                 expect(error.message).toMatch(/near \"VALUES\": syntax error.*code 1.*while compiling: INSERT INTO test_table/);
               else
-                //* expect(error.message).toMatch(/near \" \": syntax error/);
-                expect(error.message).toMatch(/incomplete input/); // XXX SQLite 3.22.0
+                expect(error.message).toMatch(/incomplete input/);
 
               // FAIL transaction & check reported transaction error:
               return true;
@@ -215,7 +228,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWindows || isWebSql || (isAndroid && isImpl2))
@@ -227,8 +240,10 @@ var mytests = function() {
               expect(error.message).toMatch(/callback raised an exception.*or.*error callback did not return false/);
             else if (isWindows)
               expect(error.message).toMatch(/error callback did not return false.*Error preparing an SQLite statement/);
-            //* else //* XXX TBD
-            //*   expect(error.message).toMatch(/error callback did not return false.*syntax error/);
+            else if (isAndroid && isImpl2)
+              expect(error.message).toMatch(/error callback did not return false.*syntax error/);
+            else
+              expect(error.message).toMatch(/error callback did not return false.*incomplete input/);
 
             isWebSql ? done() : db.close(done, done);
           }, function() {
@@ -239,8 +254,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'constraint violation', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           var db = openDatabase("Constraint-violation-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -271,21 +284,23 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
-              if (isWebSql && !isAndroid)
-                expect(true).toBe(true); // SKIP for iOS (WebKit) Web SQL
+              if (isWebSql && (!isAndroid || /Android 4.[1-3]/.test(navigator.userAgent)))
+                expect(true).toBe(true); // SKIP for iOS (WebKit) & Android 4.1-4.3 (WebKit) Web SQL
               else if (isWindows)
                 expect(error.code).toBe(0);
               else
                 expect(error.code).toBe(6);
 
               // (WebKit) Web SQL (Android/iOS) possibly with a missing 'r'
-              if (isWebSql && isAndroid)
+              if (isWebSql && /Android 4.[1-3]/.test(navigator.userAgent))
+                expect(error.message).toMatch(/column data is not unique/);
+              else if (isWebSql && isAndroid)
                 expect(error.message).toMatch(/could not execute statement due to a constr?aint failure.*19.*constraint failed/);
               else if (isWebSql)
-                expect(error.message).toMatch(/constr?aint fail/);
+                expect(error.message).toMatch(/constr?aint fail/); // [possibly missing letter on iOS (WebKit) Web SQL]
               else if (isWindows)
                 expect(error.message).toMatch(/SQLite3 step error result code: 1/);
               else if (isAndroid && !isImpl2)
@@ -307,7 +322,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWindows || isWebSql)
@@ -331,8 +346,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'SELECT uper("Test") (misspelled function name) [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           var db = openDatabase("Misspelled-function-name-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -353,7 +366,7 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
               if (isWindows || (isAndroid && isImpl2))
@@ -362,13 +375,13 @@ var mytests = function() {
                 expect(error.code).toBe(5);
 
               // ACTUAL WebKit Web SQL vs plugin error.message
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 no such function: uper/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*no such function: uper/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/no such function: uper.*code 1/);
               else
                 expect(error.message).toMatch(/no such function: uper/);
@@ -385,7 +398,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
@@ -409,8 +422,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'SELECT FROM bogus table (other database error) [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           var db = openDatabase("SELECT-FROM-bogus-table-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -432,7 +443,7 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
               if (isWindows || (isAndroid && isImpl2))
@@ -440,13 +451,13 @@ var mytests = function() {
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 no such table: BogusTable/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*no such table: BogusTable/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/no such table: BogusTable.*code 1/);
               else
                 expect(error.message).toMatch(/no such table: BogusTable/);
@@ -462,7 +473,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
@@ -486,8 +497,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'INSERT missing column [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           var db = openDatabase("INSERT-missing-column-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -510,7 +519,7 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
               if (isWindows || (isAndroid && isImpl2))
@@ -518,13 +527,13 @@ var mytests = function() {
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 table test_table has 2 columns but 1 values were supplied/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*table test_table has 2 columns but 1 values were supplied/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/table test_table has 2 columns but 1 values were supplied.*code 1.*while compiling: INSERT INTO test_table/);
               else
                 expect(error.message).toMatch(/table test_table has 2 columns but 1 values were supplied/);
@@ -541,7 +550,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
@@ -565,8 +574,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'INSERT wrong column name [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           var db = openDatabase("INSERT-wrong-column-name-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -589,7 +596,7 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
               if (isWindows || (isAndroid && isImpl2))
@@ -597,13 +604,13 @@ var mytests = function() {
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql)
+              if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*1 table test_table has no column named wrong_column/);
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
-              else if (isAndroid && !isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
                 expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*table test_table has no column named wrong_column/);
-              else if (isAndroid && isImpl2)
+              else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/table test_table has no column named wrong_column.*code 1.*while compiling: INSERT INTO test_table/);
               else
                 expect(error.message).toMatch(/table test_table has no column named wrong_column/);
@@ -620,7 +627,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
@@ -647,8 +654,6 @@ var mytests = function() {
         // claims to detect the error at the "prepare statement" stage while the
         // plugin detects the error at the "execute statement" stage.
         it(suiteName + 'CREATE VIRTUAL TABLE USING bogus module (other database error) [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           var db = openDatabase("create-virtual-table-using-bogus-module-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
           expect(db).toBeDefined();
 
@@ -670,7 +675,7 @@ var mytests = function() {
 
               // error.hasOwnProperty('message') apparently NOT WORKING on
               // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-              if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+              if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
                 expect(error.hasOwnProperty('message')).toBe(true);
 
               if (isWindows || (!isWebSql && isAndroid && isImpl2))
@@ -678,9 +683,13 @@ var mytests = function() {
               else
                 expect(error.code).toBe(5);
 
-              if (isWebSql && isAndroid)
+              if (isWebSql && isAndroid && !(/Android 4.[1-3]/.test(navigator.userAgent)))
                 expect(error.message).toMatch(/could not prepare statement.*not authorized/);
-              else if (isWebSql) // (iOS)
+              else if (isWebSql && isAndroid)
+                expect(error.message).toMatch(/not authorized/);
+              else if (isWebSql && (isBrowser && (/Chrome/.test(navigator.userAgent))))
+                expect(error.message).toMatch(/could not prepare statement.*23 not authorized/);
+              else if (isWebSql) // [iOS (WebKit) Web SQL]
                 expect(error.message).toMatch(/could not prepare statement.*1 not authorized/);
               else if (isWindows)
                 expect(error.message).toMatch(/SQLite3 step error result code: 1/);
@@ -702,7 +711,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWebSql || isWindows || (isAndroid && isImpl2))
@@ -728,8 +737,6 @@ var mytests = function() {
         // TESTS with no SQL error handler:
 
         it(suiteName + 'transaction.executeSql syntax error (command with misspelling) with no SQL error handler', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           db = openDatabase('tx-sql-syntax-error-with-no-sql-error-handler-test.db');
           db.transaction(function(transaction) {
             transaction.executeSql('SLCT 1');
@@ -741,7 +748,7 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
             if (isWindows || (isAndroid && isImpl2))
@@ -749,16 +756,18 @@ var mytests = function() {
             else
               expect(error.code).toBe(5);
 
-            if (isWebSql)
+            if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
               expect(error.message).toMatch(/could not prepare statement.*1 near \"SLCT\": syntax error/);
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
-            else if (isAndroid && !isImpl2)
+            else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
               expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*near \"SLCT\": syntax error/);
-            else if (isAndroid && isImpl2)
+            else if (!isWebSql && !isWindows && isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: near \"SLCT\": syntax error.*code 1.*while compiling: SLCT 1/);
-            else
+            else if (!isWebSql) // [iOS/macOS plugin]
               expect(error.message).toMatch(/a statement with no error handler failed.*near \"SLCT\": syntax error/);
+            else
+              expect(error.message).toMatch(/near \"SLCT\": syntax error/);
 
             isWebSql ? done() : db.close(done, done);
           }, function() {
@@ -769,8 +778,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'transaction.executeSql constraint violation with no SQL error handler', function(done) {
-          if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
-
           var db = openDatabase("Constraint-violation-with-no-sql-error-handler.db", "1.0", "Demo", DEFAULT_SIZE);
 
           db.transaction(function(tx) {
@@ -788,18 +795,20 @@ var mytests = function() {
 
             // error.hasOwnProperty('message') apparently NOT WORKING on
             // WebKit Web SQL on Android 5.x/... or iOS 10.x/...:
-            if (!isWebSql || isWindows || (isAndroid && (/Android [1-4]/.test(navigator.userAgent))))
+            if (!isWebSql || isWindows || (isAndroid && (/Android 4/.test(navigator.userAgent))))
               expect(error.hasOwnProperty('message')).toBe(true);
 
-            if (isWebSql && !isAndroid)
-              expect(true).toBe(true); // SKIP for iOS (WebKit) Web SQL
+            if (isWebSql && (!isAndroid || /Android 4.[1-3]/.test(navigator.userAgent)))
+              expect(true).toBe(true); // SKIP for iOS (WebKit) & Android 4.1-4.3 (WebKit) Web SQL
             else if (isWindows)
               expect(error.code).toBe(0);
             else
               expect(error.code).toBe(6);
 
             // (WebKit) Web SQL (Android/iOS) possibly with a missing 'r'
-            if (isWebSql && isAndroid)
+            if (isWebSql && /Android 4.[1-3]/.test(navigator.userAgent))
+              expect(error.message).toMatch(/column data is not unique/);
+            else if (isWebSql && isAndroid)
               expect(error.message).toMatch(/could not execute statement due to a constr?aint failure.*19.*constraint failed/);
             else if (isWebSql)
               expect(error.message).toMatch(/constr?aint fail/);