+++ plugins/uk.co.workingedge.phonegap.plugin.launchnavigator/src/ios/LaunchNavigator.h
@@ -53,7 +53,9 @@ typedef NS_ENUM(NSUInteger, LNApp) {
 	LNAppLyft,           // Lyft
     LNAppMapsMe,          // MAPS.ME
     LNAppCabify,          // Cabify
-    LNAppBaidu           // Baidu
+    LNAppBaidu,           // Baidu
+	LNAppTaxis99,           // 99 Taxi
+	LNAppGaode           // Gaode (Amap)
 };
 
 static NSString*const LOG_TAG = @"LaunchNavigator[native]";