+++ src/app/app.module.ts
@@ -50,10 +50,8 @@ import { Facebook } from '@ionic-native/facebook';
 import { ToggleIconButtonComponent } from '../components/toggle-icon-button/toggle-icon-button.component';
 import { SwipeVertical } from '../services/swipe-listener.service';
 import { CallNumber } from '@ionic-native/call-number';
-import { Vibration } from '@ionic-native/vibration';
 import { ConsumerContainerComponent } from '../components/consumer-container/consumer-container.component';
 import { LocationCardsService } from '../services/location-cards.service';
-import { BackgroundGeolocation } from '@ionic-native/background-geolocation';
 import { SplashScreen } from '@ionic-native/splash-screen';
 
 export class MyHammerConfig extends HammerGestureConfig {
@@ -161,9 +159,7 @@ export class MyHammerConfig extends HammerGestureConfig {
     Facebook,
     CallNumber,
     FirebaseEnvironmentService,
-    Vibration,
     LocationCardsService,
-    BackgroundGeolocation,
     SplashScreen,
   ],
 })